X and Y input vectors passed to pcolor must be casted to type Double to be compatible with Matlab R2016a.

+++ Graph/pcolorMooring2DVar.m
@@ -199,7 +199,7 @@ for i=1:lenSampleData
             dataVar(~iGoodHeight) = [];
             dataVar = reshape(dataVar, nGoodTime, nGoodHeight);
             
-            hPcolorVar(i) = pcolor(hAxMooringVar, xPcolor, yPcolor, double(dataVar'));
+            hPcolorVar(i) = pcolor(hAxMooringVar, double(xPcolor), double(yPcolor), double(dataVar'));
             set(hPcolorVar(i), 'FaceColor', 'flat', 'EdgeColor', 'none');
 
             % Let's redefine properties after pcolor to make sure grid lines appear
